.metadata ignored

+++ MCU1/stm32f401re_drivers/drivers/inc/stm32f401re_gpio_driver.h
@@ -132,6 +132,6 @@ void GPIO_TogglePin(GPIO_RegDef_t *pGPIOx, uint8_t PinNumber);
  * */
 void GPIO_IRQHandling(uint8_t PinNumber);
 void GPIO_IRQInterruptConfig(uint8_t IRQNumber, uint8_t EnorDi);
-void GPIO_IRQPRiorityConfig(uint8_t IRQNumber, uint8_t IRQPriority);
+void GPIO_IRQPRiorityConfig(uint8_t IRQNumber, uint32_t IRQPriority);
 
 #endif /* INC_STM32F401RE_GPIO_DRIVER_H_ */
